Segregated S3service into parital classes for better readability.
Implemented UploadFile method
