+++ AWS.Wrappers.S3.Tests/Files/Upload-File-1.txt
@@ -1,0 +1,1 @@
+Test
