+++ AWS.Wrappers.S3/Services/S3Service.cs
@@ -1,6 +1,7 @@
 ﻿using Amazon;
 using Amazon.S3;
 using Amazon.S3.Model;
+using Amazon.S3.Transfer;
 using Amazon.S3.Util;
 using AWS.Wrappers.S3.Interfaces;
 using AWS.Wrappers.S3.Models;
@@ -12,9 +13,8 @@ using System.Threading.Tasks;
 
 namespace AWS.Wrappers.S3.Services;
 
-public class S3Service : IS3Service
+public partial class S3Service : IS3Service
 {
-
     private readonly IAmazonS3 _s3Client;
     private readonly string _s3Region;
     #region Constructors
@@ -41,505 +41,6 @@ public class S3Service : IS3Service
 
     #endregion
 
-    #region Public Methods
-    public void BulkDownloadToLocal(S3Download[] s3Download)
-    {
-        try
-        {
-            
-        }
-        catch (Exception e)
-        {
-
-            throw;
-        }
-    }
-
-    public void BulkUpload(S3Upload[] s3Uploads)
-    {
-        throw new NotImplementedException();
-    }
-
-    public void BulkUpload(S3UploadByBytes[] s3Uploads)
-    {
-        throw new NotImplementedException();
-    }
-
-    public async Task CreateBucketAsync(string bucketName, CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            var respone = await _s3Client.PutBucketAsync(bucketName, cancellationToken);
-
-            if (respone.HttpStatusCode != System.Net.HttpStatusCode.OK)
-            {
-                throw new ArgumentException(respone.ResponseMetadata.ToString());
-            }
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public void CreateBucket(string bucketName)
-    {
-        try
-        {
-            var response = _s3Client.PutBucketAsync(bucketName).GetAwaiter().GetResult();
-
-            if (response.HttpStatusCode != System.Net.HttpStatusCode.OK)
-            {
-                throw new ArgumentException(response.ResponseMetadata.ToString());
-            }
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public async Task CreateDirectoryAsync(string bucketName, string path, CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            path = PathFormat(path);
-            
-            if (await IsDirectoryExistsAsync(bucketName, path)) return;
-
-            var request = new PutObjectRequest()
-            {
-                BucketName = bucketName,
-                StorageClass = S3StorageClass.Standard,
-                ServerSideEncryptionMethod = ServerSideEncryptionMethod.None,
-                Key = path,
-                ContentBody = string.Empty
-            };
-
-            var response = await _s3Client.PutObjectAsync(request, cancellationToken);
-
-            if (response.HttpStatusCode != System.Net.HttpStatusCode.OK)
-            {
-                throw new ArgumentException(response.ResponseMetadata.ToString());
-            }
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public void CreateDirectory(string bucketName, string path)
-    {
-        try
-        {
-            path = PathFormat(path);
-
-            if (IsDirectoryExists(bucketName, path)) return;
-
-            var request = new PutObjectRequest()
-            {
-                BucketName = bucketName,
-                StorageClass = S3StorageClass.Standard,
-                ServerSideEncryptionMethod = ServerSideEncryptionMethod.None,
-                Key = path,
-                ContentBody = string.Empty
-            };
-
-            var response = _s3Client.PutObjectAsync(request).GetAwaiter().GetResult();
-
-            if (response.HttpStatusCode != System.Net.HttpStatusCode.OK)
-            {
-                throw new ArgumentException(response.ResponseMetadata.ToString());
-            }
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public async Task DeleteBucketAsync(string bucketName, CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            if (await IsBucketExistsAsync(bucketName))
-            {
-                var response = await _s3Client.DeleteBucketAsync(bucketName, cancellationToken);
-                if (response.HttpStatusCode != System.Net.HttpStatusCode.OK)
-                {
-                    throw new ArgumentException(response.ResponseMetadata.ToString());
-                }
-                return;
-            }
-
-            throw new ArgumentException("Bucket not found");
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public void DeleteBucket(string bucketName)
-    {
-        try
-        {
-            if (IsBucketExists(bucketName))
-            {
-                var response = _s3Client.DeleteBucketAsync(bucketName).GetAwaiter().GetResult();
-
-                if (response.HttpStatusCode != System.Net.HttpStatusCode.OK)
-                {
-                    throw new ArgumentException(response.ResponseMetadata.ToString());
-                }
-                return;
-            }
-
-            throw new ArgumentException("Bucket not found");
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public async Task DeleteDirectoryAsync(string bucketName, string path, bool recursive = false, CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            path = PathFormat(path);
-
-            if (! await IsDirectoryExistsAsync(bucketName, path, cancellationToken))
-            {
-                return;
-            }
-
-            var subFolders = await GetSubDirectoriesAsync(bucketName, path, cancellationToken);
-            //Todo: Need to modify after implementing GetFiles
-            //var files = GetFiles(bucketName, path);
-
-            //if (!recursive && subFolders?.Count > 0 || files?.Count > 0)
-            if (!recursive && subFolders?.Count > 0)
-            {
-                throw new InvalidOperationException("The directory is not empty. Try recrusive = true to delete recrusively.");
-            }
-
-            DeleteObjectRequest deleteObject = new DeleteObjectRequest
-            {
-                BucketName = bucketName,
-                Key = path,
-            };
-
-            var response = await _s3Client.DeleteObjectAsync(deleteObject, cancellationToken);
-
-            if (response.HttpStatusCode != System.Net.HttpStatusCode.NoContent || response.HttpStatusCode != System.Net.HttpStatusCode.OK)
-            {
-                throw new ArgumentException(response.ResponseMetadata.ToString());
-            }
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public void DeleteDirectory(string bucketName, string path, bool recursive = false)
-    {
-        try
-        {
-            path = PathFormat(path);
-
-            if (!IsDirectoryExists(bucketName, path))
-            {
-                return;
-            }
-
-            var subFolders = GetSubDirectories(bucketName, path);
-            //Todo: Need to modify after implementing GetFiles
-            //var files = GetFiles(bucketName, path);
-
-            //if (!recursive && subFolders?.Count > 0 || files?.Count > 0)
-            if (!recursive && subFolders?.Count > 0)
-            {
-                throw new InvalidOperationException("The directory is not empty. Try recrusive = true to delete recrusively.");
-            }
-
-            DeleteObjectRequest deleteObject = new DeleteObjectRequest
-            {
-                BucketName = bucketName,
-                Key = path,
-            };
-
-            var response = _s3Client.DeleteObjectAsync(deleteObject).GetAwaiter().GetResult();
-
-            if (response.HttpStatusCode != System.Net.HttpStatusCode.NoContent || response.HttpStatusCode != System.Net.HttpStatusCode.OK)
-            {
-                throw new ArgumentException(response.ResponseMetadata.ToString());
-            }
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public void DeleteFile(string s3Key, string bucketName)
-    {
-        throw new NotImplementedException();
-    }
-
-    public void DownloadFileToLocal(S3Download s3Download)
-    {
-        throw new NotImplementedException();
-    }
-
-    public byte[] GetFileBytes(S3Download s3Download)
-    {
-        throw new NotImplementedException();
-    }
-
-    public async Task<List<string>> GetAllDirectoriesRecursiveAsync(string bucketName, string path, CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            path = PathFormat(path);
-
-            ListObjectsV2Request listObjectsV2 = new ListObjectsV2Request
-            {
-                BucketName = bucketName,
-                Prefix = path,
-                StartAfter = path
-            };
-
-            var response = await _s3Client.ListObjectsV2Async(listObjectsV2, cancellationToken);
-
-            var files = new List<string>();
-
-            if (response.HttpStatusCode != System.Net.HttpStatusCode.OK)
-            {
-                throw new ArgumentException(response.ResponseMetadata.ToString());
-            }
-
-            foreach (var item in response.S3Objects)
-            {
-                files.Add(item.Key);
-            }
-
-            return files;
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public List<string> GetAllDirectoriesRecursive(string bucketName, string path)
-    {
-        try
-        {
-            path = PathFormat(path);
-
-            ListObjectsV2Request listObjectsV2 = new ListObjectsV2Request
-            {
-                BucketName = bucketName,
-                Prefix = path,
-                StartAfter = path
-            };
-
-            var response = _s3Client.ListObjectsV2Async(listObjectsV2).GetAwaiter().GetResult();
-
-            var files = new List<string>();
-
-            if (response.HttpStatusCode != System.Net.HttpStatusCode.OK)
-            {
-                throw new ArgumentException(response.ResponseMetadata.ToString());
-            }
-
-            foreach (var item in response.S3Objects)
-            {
-                files.Add(item.Key);
-            }
-
-            return files;
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public async Task<List<string>> GetSubDirectoriesAsync(string bucketName, string path, CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            path = PathFormat(path);
-
-            ListObjectsV2Request listObjectsV2 = new ListObjectsV2Request
-            {
-                BucketName = bucketName,
-                Prefix = path,
-                StartAfter = path
-            };
-
-            var response = await _s3Client.ListObjectsV2Async(listObjectsV2, cancellationToken);
-
-            var files = new List<string>();
-
-            if (response.HttpStatusCode != System.Net.HttpStatusCode.OK)
-            {
-                throw new ArgumentException(response.ResponseMetadata.ToString());
-            }
-
-            var temp = "";
-            foreach (var item in response.S3Objects)
-            {
-                temp = item.Key.Replace(path, "");
-                if (item.Size == 0)
-                {
-                    temp = temp.Split('/')[0];
-                    if (!files.Contains(temp))
-                    {
-                        files.Add(temp);
-                    }
-                }
-            }
-
-            return files;
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public List<string> GetSubDirectories(string bucketName, string path)
-    {
-        try
-        {
-            path = PathFormat(path);
-
-            ListObjectsV2Request listObjectsV2 = new ListObjectsV2Request
-            {
-                BucketName = bucketName,
-                Prefix = path,
-                StartAfter = path
-            };
-
-            var response = _s3Client.ListObjectsV2Async(listObjectsV2).GetAwaiter().GetResult();
-
-            var files = new List<string>();
-
-            if (response.HttpStatusCode != System.Net.HttpStatusCode.OK)
-            {
-                throw new ArgumentException(response.ResponseMetadata.ToString());
-            }
-
-            var temp = "";
-            foreach (var item in response.S3Objects)
-            {
-                temp = item.Key.Replace(path, "");
-                if (item.Size == 0)
-                {
-                    temp = temp.Split('/')[0];
-                    if (files.Contains(temp))
-                    {
-                        files.Add(temp);
-                    }
-                }
-            }
-
-            return files;
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public async Task<bool> IsBucketExistsAsync(string bucketName)
-    {
-        try
-        {
-            return await AmazonS3Util.DoesS3BucketExistV2Async(_s3Client, bucketName);
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public bool IsBucketExists(string bucketName)
-    {
-        try
-        {
-            return AmazonS3Util.DoesS3BucketExistV2Async(_s3Client, bucketName).GetAwaiter().GetResult();
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public async Task<bool> IsDirectoryExistsAsync(string bucketName, string path, CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            path = PathFormat(path);
-
-            ListObjectsV2Request request = new ListObjectsV2Request
-            {
-                BucketName = bucketName,
-                Prefix = path
-            };
-
-            var response = await _s3Client.ListObjectsV2Async(request, cancellationToken);
-            return (response != null && response.S3Objects != null && response.S3Objects.Count > 0);
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public bool IsDirectoryExists(string bucketName, string path)
-    {
-        try
-        {
-            path = PathFormat(path);
-
-            ListObjectsV2Request request = new ListObjectsV2Request
-            {
-                BucketName = bucketName,
-                Prefix = path
-            };
-
-            var response = _s3Client.ListObjectsV2Async(request).GetAwaiter().GetResult();
-            return (response != null && response.S3Objects != null && response.S3Objects.Count > 0);
-        }
-        catch
-        {
-            throw;
-        }
-    }
-
-    public bool IsFileExists(string s3Key, string bucketName)
-    {
-        throw new NotImplementedException();
-    }
-
-    public void UploadFile(S3Upload s3Upload)
-    {
-        throw new NotImplementedException();
-    }
-
-    public void UploadFile(S3UploadByBytes s3Uploads)
-    {
-        throw new NotImplementedException();
-    }
-
-    #endregion
-
     #region Private Methods
 
     private void ValidateKeyConstructor(string accessKey, string secretKey)
@@ -558,16 +59,6 @@ public class S3Service : IS3Service
         path = path.Replace("\\", "/");
         path = path.EndsWith("/") ? path : path +"/";
         return path;
-    }
-
-    public List<string> GetFiles(string bucketName, string path)
-    {
-        throw new NotImplementedException();
-    }
-
-    public Task<List<string>> GetFilesAsync(string bucketName, string path, CancellationToken cancellationToken = default)
-    {
-        throw new NotImplementedException();
     }
 
     #endregion
