+++ AWS.Wrappers.S3/Versions/S3WrapperMethodsList.md
@@ -8,7 +8,7 @@
 | DeleteBucketAsync / DeleteBucket     | Deletes the bucket if exists                                                               |
 | IsBucketExistsAsync / IsBucketExists | Check if bucket exists in S3 or not.                                                       |
 
-### Operations Related to directory/folder _(S3 Object Technically)_
+### Operations Related to directory/folder _(S3 Objects)_
 
 | Method Name                                                  | Description                                                                                                                                                                    |
 |--------------------------------------------------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
@@ -17,5 +17,12 @@
 | DeleteDirectoryAsync / DeleteDirectory                       | If a folder has files or sub-folders in it, then it won't be deleted​. We need to pass recursive = true in the argument to delete all files and folders inside the given path. |
 | GetSubDirectoriesAsync / GetSubDirectories                   | It gets the list of subdirectories from the given path.                                                                                                                        |
 | GetAllDirectoriesRecursiveAsync / GetAllDirectoriesRecursive | It gets the list of directories and sub-directories under the given path.                                                                                                      |
+
+
+### Operations Related to files _(S3 Objects)_
+
+| Method Name                                                  | Description                                                                                                                                                                    |
+|--------------------------------------------------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+| UploadFileAsync / UploadFile				                   | Used to upload local file to S3																																				|
